+++ hasher_test.go
@@ -1,7 +1,10 @@
 package ssz
 
 import (
+	"fmt"
 	"testing"
+
+	"github.com/prysmaticlabs/gohashtree"
 )
 
 func TestDepth(t *testing.T) {
@@ -52,4 +55,20 @@ func TestNextPowerOfTwo(t *testing.T) {
 			t.Fatalf("num %d, expected %d but found %d", c.Num, c.Res, next)
 		}
 	}
+}
+
+func TestHashGoHashTree(t *testing.T) {
+
+	a := make([]byte, 32)
+	b := make([]byte, 32)
+	a[0] = 1
+	b[0] = 2
+
+	buf := []byte{}
+	buf = append(buf, a...)
+	buf = append(buf, b...)
+
+	gohashtree.Hash(buf, buf)
+
+	fmt.Println(buf)
 }